LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/element-mutation-expected.txt


PASS An element with global registry should not change its registry when run append into a shadow tree with scoped registry.
FAIL An element with scoped registry should not change its registry when run append out of the shadow tree. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL An element with scoped registry should not change its registry when run append into another shadow tree with different scoped registry. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS A freshly created element should preserve global registry after append into and removal from a scoped shadow root.
PASS A freshly created element should preserve global registry when run append between scoped shadow roots.
PASS An element with global registry should not change its registry when run appendChild into a shadow tree with scoped registry.
FAIL An element with scoped registry should not change its registry when run appendChild out of the shadow tree. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL An element with scoped registry should not change its registry when run appendChild into another shadow tree with different scoped registry. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS A freshly created element should preserve global registry after appendChild into and removal from a scoped shadow root.
PASS A freshly created element should preserve global registry when run appendChild between scoped shadow roots.
PASS An element with global registry should not change its registry when run prepend into a shadow tree with scoped registry.
FAIL An element with scoped registry should not change its registry when run prepend out of the shadow tree. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
FAIL An element with scoped registry should not change its registry when run prepend into another shadow tree with different scoped registry. assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
PASS A freshly created element should preserve global registry after prepend into and removal from a scoped shadow root.
PASS A freshly created element should preserve global registry when run prepend between scoped shadow roots.
PASS Declarative shadow DOM with shadowrootcustomelementregistry attribute without registry initialized should remain null registry after adoption.
